+++ CHANGELOG.md
@@ -13,10 +13,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Minimal supported `Scarb` version is now `2.10.0` (updated from `2.9.1`)
 
+#### Fixed
+
+- [`core::testing::get_available_gas`](https://docs.starknet.io/build/corelib/core-testing-get_available_gas) now works correctly in snforge tests
+
 #### Removed
 
 - Possibility to use `#[available_gas]` with unnamed argument. Use named arguments instead, e.g. `#[available_gas(l2_gas: 5)]`.
 - The deprecated command `snforge init`. Use `snforge new` to initialize new `Forge` projects
+
+### Cast
+
+#### Added
+
+`sncast declare-from` command to declare a contract by fetching it from a different Starknet instance
 
 ## [0.49.0] - 2025-09-03